requirements.txt
python-telegram-bot==21.0
anthropic>=0.40.0
streamlit==1.35.0
apscheduler==3.10.4
python-dotenv==1.0.0
plotly==5.22.0
pandas==2.2.0
